+++ packages/extractor/src/index.ts
@@ -7,3 +7,4 @@ export * from './MulticallAggregator.js'
 export * from './TokenManager.js'
 export * from './UniV2Extractor.js'
 export * from './UniV3Extractor.js'
+export * from './PermanentCache.js'
