+++ packages/sushi/package.json
@@ -214,7 +214,6 @@
     "date-fns": "3.3.1",
     "decimal.js-light": "2.5.1",
     "seedrandom": "3.0.5",
-    "serialijse": "0.3.0",
     "tiny-invariant": "1.3.1",
     "toformat": "2.0.0",
     "tsx": "^4.7.1"
